Improved working with tags and made tags both editable and deletable; allowed some models to be hidden from certain users by using the itemclass_is_hidden field and adding a new permission token required to see those models that are hidden
git-svn-id: https://smartest-php5.repos.vscmedia.com@799 3b18a885-9e44-0410-b831-aac8df1b01cb

+++ System/Applications/Settings/SettingsAjax.class.php
@@ -40,11 +40,15 @@ class SettingsAjax extends SmartestSystemApplication{
     
     public function deleteTagById(){
         
-        $t = new SmartestTag;
-        $tag_id = (int) $this->getRequestParameter('tag_id');
+        if($this->getUser()->hasToken('delete_tags')){
         
-        if($t->find($tag_id)){
-            $t->delete();
+            $t = new SmartestTag;
+            $tag_id = (int) $this->getRequestParameter('tag_id');
+            
+            if($t->find($tag_id)){
+                $t->delete();
+            }
+        
         }
         
         exit;
